[기능 추가] 출석 보상 계산, 보상 메일 전송 및 보상 메일 아이템 정보 생성, 메일 전송 롤백

+++ Models/DAO/Game/UserData.cs
@@ -5,7 +5,7 @@ public class UserData
     public int GameUserId { get; set; }
     public int PlayerId { get; set; }
     public int UserLevel { get; set; }
-    public int AttendanceCount { get; set; }
+    public Int16 AttendanceCount { get; set; }
     public DateTime LastDate { get; set; }
 
     public Int64 Gold { get; set; }
